Added TSOP38238
The Fritzing Library of parts comes with an IR Receiver Diode from the TSOP312 series, and it turns out the pin-out is different on this device than the ones from SparkFun at https://www.sparkfun.com/products/10266. I wanted my PCB to match this part, so I edited the pin-out of the original Fritzing Part to match this IR Receiver diode.